move functions to seperate file
